// src/enums/media.ts
export enum DocumentFileTypeEnum {
  PDF = "application/pdf",
  XLS = "application/excel",
  DOC = "application/document",
  CSV = "text/csv",
  TSV = "text/tab-separated-values",
  PPT = "application/powerpoint",
  TXT = "text/plain",
}

export enum ImageFileTypeEnum {
  JPEG = "image/jpeg",
  PNG = "image/png",
  WEBP = "image/webp",
  HEIC = "image/heic",
  SVG = "image/svg+xml",
  GIF = "image/gif",
  BMP = "image/bmp",
}

export enum MediaFileTypeEnum {
  DOCUMENT = 1,
  IMAGE = 2,
  PLATFORM_DOCUMENT = 3,
}